+++ src/helper/Helpers.ps1
@@ -142,6 +142,34 @@ function Test-IsMicrosoftFile {
     return $false
 }
 
+function Test-IsSystemFolder {
+
+    [CmdletBinding()]
+    param (
+        [string] $Path
+    )
+    
+    begin {
+        $SystemPaths = @()
+    }
+    
+    process {
+        # Initialize system path list
+        if ($SystemPaths.Count -eq 0) {
+            [string[]] $SystemPaths += $env:windir
+            [string[]] $SystemPaths += Join-Path -Path "$($env:windir)" -ChildPath "System"
+            [string[]] $SystemPaths += Join-Path -Path "$($env:windir)" -ChildPath "System32"
+            [string[]] $SystemPaths += Join-Path -Path "$($env:windir)" -ChildPath "Syswow64"
+            [string[]] $SystemPaths += Join-Path -Path "$($env:windir)" -ChildPath "Sysnative"
+            [string[]] $SystemPaths += $env:ProgramFiles
+            [string[]] $SystemPaths += ${env:ProgramFiles(x86)}
+            [string[]] $SystemPaths += $env:ProgramData
+        }
+
+        $SystemPaths -contains $Path.TrimEnd('\\')
+    }
+}
+
 function Get-CurrentUserSids {
 
     [CmdletBinding()] Param()